chore(cleanup): add __init__.py for transaction package and refactor import paths
