+++ src/gui/pages/transaction/transactionpage.py
